chore: export createWorkbook/createFile methods - remove ExcelBuilder Factory class and report functions directly instead - remove unnecessary `global-excel-builder` filename since we already use Rollup to add ExcelBuilder on the window object
