+++ docs/TOC.md
@@ -6,17 +6,16 @@
 
 - [Creating Workbooks](workbook-create.md)
 - [Creating Worksheets](worksheet-create.md)
-- [Adding data to a worksheet](worksheet-add-data.md)
+- [Adding data to a Worksheet](worksheet-add-data.md)
 - [Sizing/Hiding Columns](column-sizing-collapsing.md)
 - [Setting row height/style](row-height-style.md)
 - [Fonts and Colors](fonts-and-colors.md)
 - [Formatting numbers, dates, etc](number-date-formatting.md)
 - [Alignment](alignment.md)
-- [Background fillers](background-fillers.md)
+- [Background Fillers](background-fillers.md)
 - [Formulas](formulas.md)
 - [Tables](tables.md)
 - [Theming Tables](theming-tables.md)
 - [Tables Summaries](tables-summaries.md)
-- [Adding headers and footers to a worksheet](worksheet-headers-footers.md)
-- [Using web workers for large datasets](web-workers.md)
+- [Adding Headers and Footers to a Worksheet](worksheet-headers-footers.md)
 - [Inserting images into spreadsheets](inserting-pictures.md)
