+++ docs/fonts-and-colors.md
@@ -74,7 +74,7 @@ So, to bring it all together in an example:
 ```ts
 import { ExcelBuilder } from 'excel-builder-vanilla';
 
-const artistWorkbook = new ExcelBuilder().createWorkbook();
+const artistWorkbook = createWorkbook();
 const albumList = artistWorkbook.createWorksheet({ name: 'Album List' });
 const stylesheet = artistWorkbook.getStyleSheet();
 
@@ -118,6 +118,6 @@ albumList.setColumns([{ width: 30 }, { width: 20 }, { width: 10 }]);
 
 artistWorkbook.addWorksheet(albumList);
 
-const data = new ExcelBuilder().createFile(artistWorkbook);
+const data = createExcelFile(artistWorkbook);
 downloader('Artist WB.xlsx', data);
 ```
