+++ docs/tables.md
@@ -19,7 +19,7 @@ const originalData = [
   ['Crystal Method', 'Divided By Night', 8.99],
 ];
 
-const artistWorkbook = new ExcelBuilder().createWorkbook();
+const artistWorkbook = createWorkbook();
 const albumList = artistWorkbook.createWorksheet({ name: 'Album List' });
 
 const albumTable = new Table();
@@ -35,6 +35,6 @@ artistWorkbook.addWorksheet(albumList);
 
 albumList.addTable(albumTable);
 artistWorkbook.addTable(albumTable);
-const data = new ExcelBuilder().createFile(artistWorkbook);
+const data = createExcelFile(artistWorkbook);
 downloader('Artist WB.xlsx', data);
 ```
