+++ docs/worksheet-headers-footers.md
@@ -32,7 +32,7 @@ const originalData = [
   ['Crystal Method', 'Divided By Night', 8.99],
 ];
 
-const artistWorkbook = new ExcelBuilder().createWorkbook();
+const artistWorkbook = createWorkbook();
 const albumList = artistWorkbook.createWorksheet({ name: 'Album List' });
 
 albumList.setData(originalData); // <-- Here's the important part
@@ -46,6 +46,6 @@ albumList.setHeader([
 albumList.setFooter(['Date of print: &D &T', '&A', 'Page &P of &N']);
 artistWorkbook.addWorksheet(albumList);
 
-const data = new ExcelBuilder().createFile(artistWorkbook);
+const data = createExcelFile(artistWorkbook);
 downloader('Artist WB.xlsx', data);
 ```
