+++ packages/demo/src/examples/example03.ts
@@ -1,4 +1,5 @@
-import { ExcelBuilder } from 'excel-builder-vanilla';
+import { createExcelFile, createWorkbook } from 'excel-builder-vanilla';
+
 import { buildHtmlTable, downloader } from './demoUtils';
 import './example03.scss';
 
@@ -19,16 +20,16 @@ export default class Example {
     tableContainerElm.appendChild(buildHtmlTable(this.originalData));
 
     this.exportBtnElm = document.querySelector('#export') as HTMLButtonElement;
-    this.exportBtnElm.addEventListener('click', this.createExcelStruct.bind(this));
+    this.exportBtnElm.addEventListener('click', this.startProcess.bind(this));
   }
 
   unmount() {
     // remove event listeners to avoid DOM leaks
-    this.exportBtnElm.removeEventListener('click', this.createExcelStruct.bind(this));
+    this.exportBtnElm.removeEventListener('click', this.startProcess.bind(this));
   }
 
-  createExcelStruct() {
-    const artistWorkbook = new ExcelBuilder().createWorkbook();
+  startProcess() {
+    const artistWorkbook = createWorkbook();
     const albumList = artistWorkbook.createWorksheet({ name: 'Album List' });
     // const stylesheet = artistWorkbook.getStyleSheet();
 
@@ -49,7 +50,7 @@ export default class Example {
 
     artistWorkbook.addWorksheet(albumList);
 
-    new ExcelBuilder().createFile(artistWorkbook).then(excelBlob => {
+    createExcelFile(artistWorkbook).then(excelBlob => {
       const downloadOptions = {
         filename: 'Artist WB.xlsx',
         format: 'xlsx',
