+++ packages/demo/src/examples/example06.ts
@@ -1,4 +1,5 @@
-import { ExcelBuilder } from 'excel-builder-vanilla';
+import { createExcelFile, createWorkbook } from 'excel-builder-vanilla';
+
 import { downloader } from './demoUtils';
 import './example06.scss';
 
@@ -7,16 +8,16 @@ export default class Example {
 
   mount() {
     this.exportBtnElm = document.querySelector('#export') as HTMLButtonElement;
-    this.exportBtnElm.addEventListener('click', this.createExcelStruct.bind(this));
+    this.exportBtnElm.addEventListener('click', this.startProcess.bind(this));
   }
 
   unmount() {
     // remove event listeners to avoid DOM leaks
-    this.exportBtnElm.removeEventListener('click', this.createExcelStruct.bind(this));
+    this.exportBtnElm.removeEventListener('click', this.startProcess.bind(this));
   }
 
-  createExcelStruct() {
-    const artistWorkbook = new ExcelBuilder().createWorkbook();
+  startProcess() {
+    const artistWorkbook = createWorkbook();
     const albumList = artistWorkbook.createWorksheet({ name: 'Album List' });
 
     const centerAlign = artistWorkbook.getStyleSheet().createFormat({
@@ -44,7 +45,7 @@ export default class Example {
 
     artistWorkbook.addWorksheet(albumList);
 
-    new ExcelBuilder().createFile(artistWorkbook).then(excelBlob => {
+    createExcelFile(artistWorkbook).then(excelBlob => {
       const downloadOptions = {
         filename: 'Artist WB.xlsx',
         format: 'xlsx',
