+++ packages/demo/src/examples/example07.ts
@@ -1,4 +1,5 @@
-import { ExcelBuilder } from 'excel-builder-vanilla';
+import { createExcelFile, createWorkbook } from 'excel-builder-vanilla';
+
 import { downloader } from './demoUtils';
 import './example07.scss';
 
@@ -7,16 +8,16 @@ export default class Example {
 
   mount() {
     this.exportBtnElm = document.querySelector('#export') as HTMLButtonElement;
-    this.exportBtnElm.addEventListener('click', this.createExcelStruct.bind(this));
+    this.exportBtnElm.addEventListener('click', this.startProcess.bind(this));
   }
 
   unmount() {
     // remove event listeners to avoid DOM leaks
-    this.exportBtnElm.removeEventListener('click', this.createExcelStruct.bind(this));
+    this.exportBtnElm.removeEventListener('click', this.startProcess.bind(this));
   }
 
-  createExcelStruct() {
-    const artistWorkbook = new ExcelBuilder().createWorkbook();
+  startProcess() {
+    const artistWorkbook = createWorkbook();
     const albumList = artistWorkbook.createWorksheet({ name: 'Album List' });
     const stylesheet = artistWorkbook.getStyleSheet();
 
@@ -45,26 +46,29 @@ export default class Example {
       },
     });
 
-    const originalData = [
+    const originalData: any = [
       [
         { value: 'Artist', metadata: { style: header.id } },
         { value: 'Album', metadata: { style: header.id } },
         { value: 'Price', metadata: { style: header.id } },
       ],
-      [{ value: 'Buckethead', metadata: { style: artistNameFormat.id } }, 'Albino Slug', 8.99],
-      [{ value: 'Buckethead', metadata: { style: artistNameFormat.id } }, 'Electric Tears', 13.99],
-      [{ value: 'Buckethead', metadata: { style: artistNameFormat.id } }, 'Colma', 11.34],
-      [{ value: 'Crystal Method', metadata: { style: artistNameFormat.id } }, 'Vegas', 10.54],
-      [{ value: 'Crystal Method', metadata: { style: artistNameFormat.id } }, 'Tweekend', 10.64],
-      [{ value: 'Crystal Method', metadata: { style: artistNameFormat.id } }, 'Divided By Night', 8.99],
     ];
+
+    for (let i = 0; i < 500; i++) {
+      originalData.push([{ value: 'Buckethead', metadata: { style: artistNameFormat.id } }, 'Albino Slug', 8.99]);
+      originalData.push([{ value: 'Buckethead', metadata: { style: artistNameFormat.id } }, 'Electric Tears', 13.99]);
+      originalData.push([{ value: 'Buckethead', metadata: { style: artistNameFormat.id } }, 'Colma', 11.34]);
+      originalData.push([{ value: 'Crystal Method', metadata: { style: artistNameFormat.id } }, 'Vegas', 10.54]);
+      originalData.push([{ value: 'Crystal Method', metadata: { style: artistNameFormat.id } }, 'Tweekend', 10.64]);
+      originalData.push([{ value: 'Crystal Method', metadata: { style: artistNameFormat.id } }, 'Divided By Night', 8.99]);
+    }
 
     albumList.setData(originalData);
     albumList.setColumns([{ width: 30 }, { width: 20 }, { width: 10 }]);
 
     artistWorkbook.addWorksheet(albumList);
 
-    new ExcelBuilder().createFile(artistWorkbook).then(excelBlob => {
+    createExcelFile(artistWorkbook).then(excelBlob => {
       const downloadOptions = {
         filename: 'Artist WB.xlsx',
         format: 'xlsx',
