+++ packages/demo/src/examples/example09.ts
@@ -1,4 +1,5 @@
-import { ExcelBuilder, Table } from 'excel-builder-vanilla';
+import { Table, createExcelFile, createWorkbook } from 'excel-builder-vanilla';
+
 import { downloader } from './demoUtils';
 import './example09.scss';
 
@@ -7,15 +8,15 @@ export default class Example {
 
   mount() {
     this.exportBtnElm = document.querySelector('#export') as HTMLButtonElement;
-    this.exportBtnElm.addEventListener('click', this.createExcelStruct.bind(this));
+    this.exportBtnElm.addEventListener('click', this.startProcess.bind(this));
   }
 
   unmount() {
     // remove event listeners to avoid DOM leaks
-    this.exportBtnElm.removeEventListener('click', this.createExcelStruct.bind(this));
+    this.exportBtnElm.removeEventListener('click', this.startProcess.bind(this));
   }
 
-  createExcelStruct() {
+  startProcess() {
     const originalData = [
       ['Artist', 'Album', 'Price'],
       ['Buckethead', 'Albino Slug', 8.99],
@@ -27,7 +28,7 @@ export default class Example {
     ];
 
     // require(['excel-builder.js/excel-builder', 'excel-builder.js/Excel/Table','download'], function (EB, Table, downloader) {
-    const artistWorkbook = new ExcelBuilder().createWorkbook();
+    const artistWorkbook = createWorkbook();
     const albumList = artistWorkbook.createWorksheet({ name: 'Album List' });
 
     const albumTable = new Table();
@@ -44,7 +45,7 @@ export default class Example {
     albumList.addTable(albumTable);
     artistWorkbook.addTable(albumTable);
 
-    new ExcelBuilder().createFile(artistWorkbook).then(excelBlob => {
+    createExcelFile(artistWorkbook).then(excelBlob => {
       const downloadOptions = {
         filename: 'Artist WB.xlsx',
         format: 'xlsx',
