+++ packages/demo/src/examples/example10.ts
@@ -1,4 +1,5 @@
-import { ExcelBuilder, Table } from 'excel-builder-vanilla';
+import { Table, createExcelFile, createWorkbook } from 'excel-builder-vanilla';
+
 import { downloader } from './demoUtils';
 import './example10.scss';
 
@@ -7,15 +8,15 @@ export default class Example {
 
   mount() {
     this.exportBtnElm = document.querySelector('#export') as HTMLButtonElement;
-    this.exportBtnElm.addEventListener('click', this.createExcelStruct.bind(this));
+    this.exportBtnElm.addEventListener('click', this.startProcess.bind(this));
   }
 
   unmount() {
     // remove event listeners to avoid DOM leaks
-    this.exportBtnElm.removeEventListener('click', this.createExcelStruct.bind(this));
+    this.exportBtnElm.removeEventListener('click', this.startProcess.bind(this));
   }
 
-  createExcelStruct() {
+  startProcess() {
     const originalData = [
       ['Artist', 'Album', 'Price'],
       ['Buckethead', 'Albino Slug', 8.99],
@@ -26,7 +27,7 @@ export default class Example {
       ['Crystal Method', 'Divided By Night', 8.99],
     ];
 
-    const artistWorkbook = new ExcelBuilder().createWorkbook();
+    const artistWorkbook = createWorkbook();
     const albumList = artistWorkbook.createWorksheet({ name: 'Album List' });
 
     const stylesheet = artistWorkbook.getStyleSheet();
@@ -59,7 +60,7 @@ export default class Example {
     albumList.addTable(albumTable);
     artistWorkbook.addTable(albumTable);
 
-    new ExcelBuilder().createFile(artistWorkbook).then(excelBlob => {
+    createExcelFile(artistWorkbook).then(excelBlob => {
       const downloadOptions = {
         filename: 'Artist WB.xlsx',
         format: 'xlsx',
