+++ packages/excel-builder-vanilla/src/Excel/__tests__/Util.spec.ts
@@ -1,4 +1,4 @@
-import { describe, expect, it, test } from 'vitest';
+import { describe, expect, it } from 'vitest';
 import { Util } from '../Util';
 
 describe('utility functions', () => {
