+++ packages/excel-builder-vanilla/src/index.ts
@@ -1,6 +1,5 @@
 export * from './Excel';
+export * from './factory';
 export * from './interfaces';
 export * from './utilities';
 export * from './Template';
-export * from './excel-builder';
-// export * from './global-excel-builder';
